Fixed issue with EdgeList.
In the transactional implementation, we didn't always create a
serialized edge list in the eListMap when we needed to. In particular,
when the head segment has no elements, but there are tail segments
(which can happen when the last element that was added triggered a head
segment split and the target split point was set to 0). In this case the
tail segment processing code would encounter a seg fault trying to
reference a non-existent serialized edge list in the map.

Fixed this to create serialized edge lists in the map also in the case
when there are tail segments but no elements in the head segment.