Optimized batchRead to not create zero-len byte arrays.
Previously EdgeList.batchRead() would fill in a zero length byte array
into the returned serializedEdge objects if the edge had no properties.
But this creates unneeded objects just to convey the non-existence of
something. So now we just use null. In the case where the use wants the
edge properties in TraversalResult filled in, we create empty hash maps
in TraversalResult.pMap.