Added byte array member to UInt128.
Now UInt128 stores the byte array as a member for easy use for forming
RAMCloud keys for reading edge lists. Now in the common case, byte
arrays for UInt128s are read from edge lists and then directly used for
creating RAMCloud keys.

We still maintain the upper and lower longs, however, for the same of
implementing equals() performantly, because byte comparison is super
slow.