Changed fillProperties to getProperties.
We don't use vertices any more to store properties, we store them in a separate property map.  This
is because if we store them in vertex objects we need to make sure that the vertex objects get
properly deduplicated in memory (otherwise you have two objects representing the same vertex with
possibly inconsistent property maps, or commonly one has its map filled in and the other does not).