Changed default segment size limit back to 8KB.
I had changed this from 8KB to 2KB, and this actually broke TorcDB2 with working
with 8KB edge lists because we use the DEFAULT_SEGMENT_SIZE_LIMIT to size
ByteBuffers. So we were getting a buffer overflow when we tried to read 8KB
edge lists into ByteBuffers sized for 2KB edge lists.