Added hashmapuint128keys test.
A test that measures the performance of hashmap operations when used
with UInt128 objects as keys. The reason this test exists in the first
place is because I found that the performance of get/put operations was
affected by the values of UInt128 objects. The performance of hashCode()
was measured, and found not to be the source of the performance
variations.